requirements.txt
fastapi
uvicorn[standard]
django
celery[redis]
redis
httpx
pytest
python-dotenv
